tools/crawler/requirements.txt
# Google Drive API
google-auth>=2.27.0
google-api-python-client>=2.114.0

# Async HTTP client
aiohttp>=3.9.0

# YAML processing
pyyaml>=6.0.1

# Environment variables
python-dotenv>=1.0.0

# MCP SDK (Model Context Protocol)
mcp>=1.0.0
